# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LoanFlow API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  EN_ATTENTE: "EN_ATTENTE",
  ACCEPTE: "ACCEPTE",
  REFUSE: "REFUSE",
  CONTRAT_ENVOYE: "CONTRAT_ENVOYE",
  CONTRAT_SIGNE: "CONTRAT_SIGNE",
  EN_TRAITEMENT: "EN_TRAITEMENT",
  FONDS_DISPONIBLES: "FONDS_DISPONIBLES",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  USER: "USER",
  ADMIN: "ADMIN",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  fullName: string | null;
  role: UserRole;
  walletBalance: number;
  createdAt: string;
}

export interface UploadedDocument {
  filename: string;
  contentType: string;
  dataBase64: string;
}

export interface UploadedDocumentRef {
  id: string;
  filename: string;
  contentType: string;
  uploadedAt: string;
}

export interface CreateLoanInput {
  /** @minLength 2 */
  applicantName: string;
  applicantEmail: string;
  /** @minimum 100 */
  amount: number;
  /**
   * @minimum 3
   * @maximum 120
   */
  durationMonths: number;
  /** @minimum 0 */
  monthlyIncome: number;
  /** @nullable */
  purpose?: string | null;
  documents: UploadedDocument[];
}

export interface Loan {
  id: string;
  applicantName: string;
  applicantEmail: string;
  amount: number;
  durationMonths: number;
  monthlyIncome: number;
  /** @nullable */
  purpose: string | null;
  status: LoanStatus;
  /** @nullable */
  decisionAt: string | null;
  /** @nullable */
  contractSignedAt: string | null;
  /** @nullable */
  fundsAvailableAt: string | null;
  /** @nullable */
  processingUntil: string | null;
  withdrawnAmount: number;
  availableBalance: number;
  /** @nullable */
  adminNote: string | null;
  createdAt: string;
  updatedAt: string;
}

/**
 * @nullable
 */
export type LoanDetailGeneratedContract = {
  id: string;
  filename: string;
  uploadedAt: string;
} | null;

/**
 * @nullable
 */
export type LoanDetailSignedContract = {
  id: string;
  filename: string;
  uploadedAt: string;
} | null;

export interface TimelineEvent {
  id: string;
  kind: string;
  message: string;
  createdAt: string;
}

export type LoanDetail = Loan & {
  documents: UploadedDocumentRef[];
  /** @nullable */
  generatedContract: LoanDetailGeneratedContract;
  /** @nullable */
  signedContract: LoanDetailSignedContract;
  timeline: TimelineEvent[];
};

export type AdminLoan = Loan & {
  userEmail: string;
  /** @nullable */
  userFullName: string | null;
};

export type AdminLoanDetail = LoanDetail & {
  userEmail: string;
  /** @nullable */
  userFullName: string | null;
};

export type LoanDecisionInputDecision =
  (typeof LoanDecisionInputDecision)[keyof typeof LoanDecisionInputDecision];

export const LoanDecisionInputDecision = {
  ACCEPT: "ACCEPT",
  REFUSE: "REFUSE",
} as const;

export interface LoanDecisionInput {
  decision: LoanDecisionInputDecision;
  /** @nullable */
  adminNote?: string | null;
}

export type AdvanceStatusInputAction =
  (typeof AdvanceStatusInputAction)[keyof typeof AdvanceStatusInputAction];

export const AdvanceStatusInputAction = {
  SEND_CONTRACT: "SEND_CONTRACT",
  START_PROCESSING: "START_PROCESSING",
  RELEASE_FUNDS: "RELEASE_FUNDS",
} as const;

export interface AdvanceStatusInput {
  action: AdvanceStatusInputAction;
}

export interface UploadFileInput {
  filename: string;
  contentType: string;
  dataBase64: string;
}

export interface WithdrawInput {
  /** @minimum 1 */
  amount: number;
}

export interface ActivityItem {
  id: string;
  loanId: string;
  applicantName: string;
  kind: string;
  message: string;
  createdAt: string;
}

export interface AdminStats {
  totalLoans: number;
  pending: number;
  accepted: number;
  refused: number;
  contractSent: number;
  contractSigned: number;
  processing: number;
  fundsAvailable: number;
  totalAmountRequested: number;
  totalAmountAccepted: number;
  totalDisbursed: number;
}

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Forbidden
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;