# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoanFlow API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/loans:
    get:
      operationId: adminListLoans
      tags: [admin]
      summary: Admin - list all loan applications
      responses:
        "200":
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLoan"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/loans/{id}:
    get:
      operationId: adminGetLoan
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loan detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoanDetail"
  /admin/loans/{id}/decision:
    patch:
      operationId: adminDecideLoan
      tags: [admin]
      summary: Admin - accept or refuse a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanDecisionInput"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoan"
  /admin/loans/{id}/advance:
    post:
      operationId: adminAdvanceStatus
      tags: [admin]
      summary: Admin - advance loan to next operational status (mark contract sent / start processing / release funds)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceStatusInput"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoan"
...
components:
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoanStatus:
      type: string
      enum:
        - EN_ATTENTE
        - ACCEPTE
        - REFUSE
        - CONTRAT_ENVOYE
        - CONTRAT_SIGNE
        - EN_TRAITEMENT
        - FONDS_DISPONIBLES
...
    UploadedDocumentRef:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        contentType:
          type: string
        uploadedAt:
          type: string
      required: [id, filename, contentType, uploadedAt]
...
    Loan:
      type: object
      properties:
        id:
          type: string
        applicantName:
          type: string
        applicantEmail:
          type: string
        amount:
          type: number
        durationMonths:
          type: integer
        monthlyIncome:
          type: number
        purpose:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LoanStatus"
        decisionAt:
          type: ["string", "null"]
        contractSignedAt:
          type: ["string", "null"]
        fundsAvailableAt:
          type: ["string", "null"]
        processingUntil:
          type: ["string", "null"]
        withdrawnAmount:
          type: number
        availableBalance:
          type: number
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - applicantName
        - applicantEmail
        - amount
        - durationMonths
        - monthlyIncome
        - purpose
        - status
        - decisionAt
        - contractSignedAt
        - fundsAvailableAt
        - processingUntil
        - withdrawnAmount
        - availableBalance
        - adminNote
        - createdAt
        - updatedAt
    LoanDetail:
      allOf:
        - $ref: "#/components/schemas/Loan"
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: "#/components/schemas/UploadedDocumentRef"
            generatedContract:
              type: ["object", "null"]
              properties:
                id:
                  type: string
                filename:
                  type: string
                uploadedAt:
                  type: string
              required: [id, filename, uploadedAt]
            signedContract:
              type: ["object", "null"]
              properties:
                id:
                  type: string
                filename:
                  type: string
                uploadedAt:
                  type: string
              required: [id, filename, uploadedAt]
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
          required: [documents, generatedContract, signedContract, timeline]
    AdminLoan:
      allOf:
        - $ref: "#/components/schemas/Loan"
        - type: object
          properties:
            userEmail:
              type: string
            userFullName:
              type: ["string", "null"]
          required: [userEmail, userFullName]
    AdminLoanDetail:
      allOf:
        - $ref: "#/components/schemas/LoanDetail"
        - type: object
          properties:
            userEmail:
              type: string
            userFullName:
              type: ["string", "null"]
          required: [userEmail, userFullName]
    LoanDecisionInput:
      type: object
      properties:
        decision:
          type: string
          enum: [ACCEPT, REFUSE]
        adminNote:
          type: ["string", "null"]
      required: [decision]
    AdvanceStatusInput:
      type: object
      properties:
        action:
          type: string
          enum: [SEND_CONTRACT, START_PROCESSING, RELEASE_FUNDS]
      required: [action]
...
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required: [id, kind, message, createdAt]
...