# lib/api-zod/src/generated/types/adminLoan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LoanFlow API
 * OpenAPI spec version: 0.1.0
 */
import type { Loan } from "./loan";

export type AdminLoan = Loan & {
  userEmail: string;
  /** @nullable */
  userFullName: string | null;
};