# lib/api-zod/src/generated/types/loanStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LoanFlow API
 * OpenAPI spec version: 0.1.0
 */

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  EN_ATTENTE: "EN_ATTENTE",
  ACCEPTE: "ACCEPTE",
  REFUSE: "REFUSE",
  CONTRAT_ENVOYE: "CONTRAT_ENVOYE",
  CONTRAT_SIGNE: "CONTRAT_SIGNE",
  EN_TRAITEMENT: "EN_TRAITEMENT",
  FONDS_DISPONIBLES: "FONDS_DISPONIBLES",
} as const;