# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LoanFlow API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  id: string;
  kind: string;
  message: string;
  createdAt: string;
}